Updating createCategory component
